improvements in the orms and creation of the auth system

+++ src/orms/messageorm.js
@@ -8,7 +8,7 @@ module.exports = class MessageOrm {
 		this.con = App.MysqlManager.con;
 		this.props = {
 			prefix: "MESSAGE-ORM",
-			table: "MESSAGES"
+			table: `${this.App.config.db}.MESSAGES`
 		};
 	}
 	
@@ -17,94 +17,116 @@ module.exports = class MessageOrm {
 	
 	/**
 	 * This function create a message in database
-	 * @param data {Object} data is text of the message, the chatName associated, the userId associated, and the date of the message
+	 * @param {Object} data is text of the message, the chatName associated, the userId associated, and the date of the message
+	 * @param {*} callback
 	 */
-	create(data){
-		this.con.query(`INSERT INTO ${props.table} VALUES (NULL, "${data.text}", "${data.chat}", ${data.userId}, "${data.date}"`, (err, res) => {
+	create(data, callback = undefined){
+		const query = `INSERT INTO ${this.props.table} VALUES (NULL, "${data.text}", "${data.chat}", ${data.userId}, "${data.date}"`;
+		
+		if (this.App.isNull(callback)) this.con.query(query, (err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Creating message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function delete a message in database
-	 * @param data {Object} data is the id of the message
+	 * @param {Object} data is the id of the message
+	 * @param {*} callback
 	 */
-	deleteById(data){
-		this.con.query(`DELETE FROM ${props.table} WHERE id=${data.id}`, (err, res) => {
+	deleteById(data, callback = undefined){
+		const query = `DELETE FROM ${this.props.table} WHERE id=${data.id}`;
+		
+		if (this.App.isNull(callback)) this.con.query(query, (err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Deleting message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function delete a message in database
-	 * @param data {Object} data is the username associated to message
+	 * @param {Object} data is the username associated to message
+	 * @param {*} callback
 	 */
-	deleteByUserName(data){
-		const query = `DELETE FROM $(props.table} WHERE userId=(
+	deleteByUserName(data, callback = undefined){
+		const query = `DELETE FROM $(this.props.table} WHERE userId=(
 			SELECT id FROM ${this.App.UserOrm.props.table} WHERE username="${data.username}"
 		)`;
 		
-		this.con.query(query, (err, res) => {
+		if (this.App.isNull(callback)) this.con.query(query, (err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Deleting message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function delete a message in database
-	 * @param data {Object} data is the userId associated to message
+	 * @param {Object} data is the userId associated to message
+	 * @param {*} callback
 	 */
-	deleteByUserId(data){
-		this.con.query(`DELETE FROM $(props.table} WHERE userId=${data.userId}`, (err, res) => {
+	deleteByUserId(data, callback = undefined){
+		const query = `DELETE FROM $(this.props.table} WHERE userId=${data.userId}`;
+		
+		if (this.App.isNull(callback)) this.con.query(query, (err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Deleting message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function update a message in database
-	 * @param data {Object} data is the id of the message
+	 * @param {Object} data is the id of the message
+	 * @param {*} callback
 	 */
-	updateById(data){
-		this.con.query(`UPDATE ${props.table} SET text="${data.text}" WHERE id=${data.id}`,(err, res) => {
+	updateById(data, callback = undefined){
+		const query = `UPDATE ${this.props.table} SET text="${data.text}" WHERE id=${data.id}`;
+		
+		if (this.App.isNull(callback)) this.con.query(query,(err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Updating message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function update a message in database 
-	 * @param data {Object} data is the username associated to the message
+	 * @param {Object} data is the username associated to the message
+	 * @param {*} callback
 	 */
-	updateByUserName(data){
-		const query = `UPDATE ${props.table} SET text="${datat.text}" WHERE userId=(
+	updateByUserName(data, callback = undefined){
+		const query = `UPDATE ${this.props.table} SET text="${datat.text}" WHERE userId=(
 			SELECT id FROM ${this.App.UserOrm.props.table} WHERE username="${data.username}"
 		)`;
 		
-		this.con.query(query,(err, res) => {
+		if (this.App.isNull(callback)) this.con.query(query,(err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Updating message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
 	
 	
 	/**
 	 * This function update a message in database 
-	 * @param data {Object} data is the userId associated to the message
+	 * @param {Object} data is the userId associated to the message
+	 * @param {*} callback
 	 */
-	updateByUserId(data){
-		this.con.query(`UPDATE ${props.table} SET text="${data.text}" WHERE userId=${data.userId}`,(err, res) => {
+	updateByUserId(data, callback = undefined){
+		const query = `UPDATE ${this.props.table} SET text="${data.text}" WHERE userId=${data.userId}`;
+		
+		if (this.App.isNull(callback)) this.con.query(query,(err, res) => {
 			if (err) this.App.throwErr(err);
 			else this.App.debug(`Updating message with: ${data}`, this.props.prefix)
-		})
+		});
+		else this.con.query(query, callback)
 	}
-	
-	
 }
